addressed what I could
