+++ app/src/main/java/bg/zahov/app/ui/history/info/HistoryInfoViewModel.kt
@@ -9,46 +9,11 @@ import bg.zahov.app.data.model.ToastManager
 import bg.zahov.app.data.model.Workout
 import bg.zahov.app.data.model.WorkoutState
 import bg.zahov.app.data.provider.WorkoutStateManager
+import bg.zahov.app.data.provider.model.HistoryInfoWorkout
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-
-/**
- * Represents detailed information about a specific exercise within a workout.
- *
- * @property exerciseName The name of the exercise.
- * @property sets A list of sets performed for the exercise, represented as strings.
- * @property oneRepMaxes A list of one-rep maxes for the exercise, represented as strings.
- */
-data class ExerciseDetails(
-    val exerciseName: String,
-    val sets: List<String>,
-    val oneRepMaxes: List<String>,
-)
-
-/**
- * Represents the detailed information of a workout history entry.
- *
- * @property id The unique identifier of the workout.
- * @property workoutName The name of the workout.
- * @property workoutDate The date when the workout was performed.
- * @property duration The duration of the workout, represented as a string.
- * @property volume The total volume lifted during the workout, represented as a string.
- * @property prs The personal records achieved during the workout, represented as a string.
- * @property exercisesInfo A list of detailed information for each exercise in the workout.
- * @property isDeleted indicates whether the workout has been deleted
- */
-data class HistoryInfoWorkout(
-    val id: String = "",
-    val workoutName: String = "",
-    val workoutDate: String = "",
-    val duration: String = "",
-    val volume: String = "",
-    val prs: String = "",
-    val exercisesInfo: List<ExerciseDetails> = listOf(),
-    val isDeleted: Boolean = false,
-)
 
 /**
  * ViewModel for managing the state and actions related to workout history details.
@@ -93,7 +58,7 @@ class HistoryInfoViewModel(
                 }
             }
             launch {
-                workoutProvider.getClickedHistoryWorkout().collect {
+                workoutProvider.clickedPastWorkout.collect {
                     _uiState.update { old ->
                         old.copy(
                             id = it.id,
@@ -135,7 +100,6 @@ class HistoryInfoViewModel(
      * Displays a toast notification if the workout already exists as a template.
      */
     fun saveAsTemplate() {
-
         if (templates.any { it.id == getCorrespondingWorkout()?.id }) {
             toastManager.showToast(R.string.workout_exists_toast)
             return
